Move specs to the right folder and delete files that are not used anymore
