+++ packages/yoastseo/spec/languageProcessing/helpers/prominentWords/determineProminentWordsEnglishSpec.js
