+++ packages/yoastseo/spec/languageProcessing/helpers/prominentWords/determineProminentWordsFrenchSpec.js
