+++ packages/yoastseo/spec/languageProcessing/helpers/prominentWords/determineProminentWordsGermanSpec.js
