+++ packages/yoastseo/spec/languageProcessing/helpers/prominentWords/determineProminentWordsPortugueseSpec.js
