+++ packages/yoastseo/spec/languageProcessing/helpers/prominentWords/determineProminentWordsSwedishSpec.js
