+++ packages/yoastseo/spec/scoring/helpers/assessments/checkForTooLongSentencesSpec.js
@@ -1,4 +1,4 @@
-import checkForTooLongSentences from "../../src/scoring/helpers/assessments/checkForTooLongSentences.js";
+import checkForTooLongSentences from "../../../../src/scoring/helpers/assessments/checkForTooLongSentences.js";
 
 describe( "Checks if sentences are too long", function() {
 	it( "Returns no sentences, none are too long", function() {
