+++ packages/yoastseo/spec/scoring/helpers/assessments/recommendedKeywordCountSpec.js
@@ -1,4 +1,4 @@
-import recommendedKeywordCount from "../../src/scoring/helpers/assessments/recommendedKeywordCount.js";
+import recommendedKeywordCount from "../../../../src/scoring/helpers/assessments/recommendedKeywordCount.js";
 const maxRecommendedDensity = 3;
 const minRecommendedDensity = 0.5;
