+++ packages/yoastseo/spec/scoring/interpreters/scoreToRatingSpec.js
@@ -1,4 +1,4 @@
-import scoreToRating from "../../src/scoring/interpreters/scoreToRating.js";
+import scoreToRating from "../../../src/scoring/interpreters/scoreToRating.js";
 
 describe( "An interpreter that turns numeric scores into ratings", function() {
 	it( "returns 'bad' for a score lower than or equal to 4", function() {
