+++ packages/yoastseo/spec/scoring/renderers/AssessorPresenterSpec.js
@@ -1,6 +1,6 @@
-import AssessorPresenter from "../../src/scoring/renderers/AssessorPresenter.js";
-import AssessmentResult from "../../src/values/AssessmentResult";
-import Factory from "../specHelpers/factory.js";
+import AssessorPresenter from "../../../src/scoring/renderers/AssessorPresenter.js";
+import AssessmentResult from "../../../src/values/AssessmentResult";
+import Factory from "../../specHelpers/factory.js";
 
 describe( "an AssessorPresenter", function() {
 	let i18n;
