+++ packages/yoastseo/spec/values/PaperSpec.js
